added view adding ,search optimization and removed bug while deleting files

+++ admin/production/deletefile.php
@@ -1,17 +1,22 @@
 <?php
 require('../../connection/config.php');
+if(isset($_GET['id'])){
 $id = $_GET['id'];
+echo $id;
 
 $sql2 = "SELECT * FROM videos WHERE id = '$id'";
 $result2 = mysqli_query($conn, $sql2);
-$row = $result2->fetch_assoc();
-$path = 'Uploads/videos/' . $row['Episode_Name'] . "." . $row['ext'];
+$row = mysqli_fetch_array($result2);
+
+$path = '../../Uploads/videos/'.$row['A_Name'] .'/'. $row['Ep_Video'] ;
 $unlink = unlink($path);
+if($unlink){  
+    $sql = "DELETE FROM videos where id='$id'";
+    $result = mysqli_query($conn, $sql);
+}
 
-$sql = "DELETE FROM videos where id='$id'";
-$result = mysqli_query($conn, $sql);
 
-
-if ($result && $unlink) {
-    header("location:managefile.php");
+ if ($result && $unlink) {
+     header("location:managefile.php");
+ }
 }
